--- OASMan_ESP32/src/ble.h
@@ -1,4 +1,0 @@
-#ifndef ble_h
-#define ble_h
-
-#endif
