+++ OASMan_ESP32/src/bluetooth/ble.h
@@ -1,0 +1,6 @@
+#ifndef ble_h
+#define ble_h
+
+// Tutorial: https://www.youtube.com/watch?v=s3yoZa6kzus&ab_channel=MoThunderz
+
+#endif
