+++ OASMan_ESP32/src/components/solenoid.h
